+++ Pandora/Paradigm/Structure/Concrete.hs
@@ -1,0 +1,3 @@
+module Pandora.Paradigm.Structure.Concrete (module Exports) where
+
+import Pandora.Paradigm.Structure.Concrete.Stack as Exports