+++ Pandora/Paradigm/Structure/Property.hs
@@ -1,0 +1,4 @@
+module Pandora.Paradigm.Structure.Property (module Exports) where
+
+import Pandora.Paradigm.Structure.Property.Nonempty as Exports
+import Pandora.Paradigm.Structure.Property.Hollow as Exports